Damho

+++ src/main/java/com/nhnacademy/midoo/gateway/domain/CommentPutRequest.java
@@ -4,8 +4,8 @@ import lombok.Value;
 
 @Value
 public class CommentPutRequest {
-    int commentId;
+    long commentId;
     String commentContent;
-    int taskId;
-    int accountId;
+    long taskId;
+    long accountId;
 }
